// src/types/project.ts
export type ProjectStatus = "התחיל" | "בתהליך עבודה" | "הושלם";

export type Project = {
  id: string;
  createdAt?: string;
  updatedAt?: string;
  projectName: string;
  clientId: string;
  client?: { id: string; clientName: string; businessName: string };
  status: ProjectStatus;
  totalAmount: number;
  paidAmount: number;
  notes?: string;
};